Moved code specific to method 2 into its own module.
